// frontend/src/gen/product_connect.ts
// @generated by protoc-gen-connect-es v1.7.0 with parameter "target=ts,import_extension=none"
// @generated from file product.proto (package product, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { CreateProductRequest, CreateToppingRequest, DeleteProductRequest, DeleteToppingRequest, GetProductRequest, ListProductsRequest, ListProductsResponse, ListToppingsResponse, Product, Topping, UpdateProductRequest } from "./product_pb";
import { MethodKind } from "@bufbuild/protobuf";
import { PaginationRequest, SuccessResponse } from "./common_pb";

/**
 * @generated from service product.ProductService
 */
export const ProductService = {
  typeName: "product.ProductService",
  methods: {
    /**
     * @generated from rpc product.ProductService.ListProducts
     */
    listProducts: {
      name: "ListProducts",
      I: ListProductsRequest,
      O: ListProductsResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc product.ProductService.GetProduct
     */
    getProduct: {
      name: "GetProduct",
      I: GetProductRequest,
      O: Product,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc product.ProductService.CreateProduct
     */
    createProduct: {
      name: "CreateProduct",
      I: CreateProductRequest,
      O: Product,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc product.ProductService.UpdateProduct
     */
    updateProduct: {
      name: "UpdateProduct",
      I: UpdateProductRequest,
      O: Product,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc product.ProductService.DeleteProduct
     */
    deleteProduct: {
      name: "DeleteProduct",
      I: DeleteProductRequest,
      O: SuccessResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc product.ProductService.ListToppings
     */
    listToppings: {
      name: "ListToppings",
      I: PaginationRequest,
      O: ListToppingsResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc product.ProductService.CreateTopping
     */
    createTopping: {
      name: "CreateTopping",
      I: CreateToppingRequest,
      O: Topping,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc product.ProductService.DeleteTopping
     */
    deleteTopping: {
      name: "DeleteTopping",
      I: DeleteToppingRequest,
      O: SuccessResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;