// frontend/src/gen/category_connect.ts
// @generated by protoc-gen-connect-es v1.7.0 with parameter "target=ts,import_extension=none"
// @generated from file category.proto (package category, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { PaginationRequest, SuccessResponse } from "./common_pb";
import { Category, CreateCategoryRequest, DeleteCategoryRequest, GetCategoryRequest, ListCategoriesResponse, UpdateCategoryRequest } from "./category_pb";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service category.CategoryService
 */
export const CategoryService = {
  typeName: "category.CategoryService",
  methods: {
    /**
     * @generated from rpc category.CategoryService.ListCategories
     */
    listCategories: {
      name: "ListCategories",
      I: PaginationRequest,
      O: ListCategoriesResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc category.CategoryService.GetCategory
     */
    getCategory: {
      name: "GetCategory",
      I: GetCategoryRequest,
      O: Category,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc category.CategoryService.CreateCategory
     */
    createCategory: {
      name: "CreateCategory",
      I: CreateCategoryRequest,
      O: Category,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc category.CategoryService.UpdateCategory
     */
    updateCategory: {
      name: "UpdateCategory",
      I: UpdateCategoryRequest,
      O: Category,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc category.CategoryService.DeleteCategory
     */
    deleteCategory: {
      name: "DeleteCategory",
      I: DeleteCategoryRequest,
      O: SuccessResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;